Group bosk-mongo test classes by purpose.
Merge the no-op write tests into MongoDriverNoOpWriteTest, whose nested
ScalarWrite and RecordWrite classes keep the two different graft-point
configurations; merge the load and refurbish race tests into MongoDriverRaceTest;
group the three BsonFormatter test classes into nested classes of one
BsonFormatterTest; and rename the orphaned MiscMongoDriverTest to ManifestTest,
which states its purpose. No test scenarios are lost.
