+++ bosk-mongo/src/test/java/works/bosk/drivers/mongo/internal/ManifestTest.java
@@ -5,7 +5,12 @@ import works.bosk.exceptions.InvalidTypeException;
 
 import static works.bosk.TypeValidation.validateType;
 
-public class MiscMongoDriverTest {
+/**
+ * Tests of {@link Manifest}, the document that records which format and format
+ * version a collection is stored in. This guards the manifest's own structure,
+ * which must remain a valid bosk state type as the driver evolves.
+ */
+public class ManifestTest {
 	@Test
 	void manifest_passesTypeValidation() throws InvalidTypeException {
 		validateType(Manifest.class);
